Refactor books manager
